Initialize React on Raspberry Pi project
This commit sets up the initial structure for a React application that can be containerized using Docker and deployed onto a Raspberry Pi. The setup includes a set of scripts in the package.json file for building the Docker image, running the Docker container, and deploying the Docker image to a Raspberry Pi. This commit serves as the main structure of the project and likely won't undergo significant changes. Future enhancements and updates will be managed in separate branches and merged into this main branch as needed.